requirements.txt
google-genai
gspread
google-api-python-client
google-auth
streamlit
openpyxl
python-docx
